requirements.txt
ai2thor==5.0.0
ultralytics>=8.3.0
opencv-python>=4.9.0
numpy>=1.26.0
python-dotenv>=1.0.1
pillow>=10.0.0
transformers>=4.45.0
accelerate>=0.33.0
sentencepiece>=0.2.0
